+++ app/src/main/java/net/kuludu/smartrecite/WrongActivity.java
@@ -4,15 +4,24 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.SharedPreferences;
 import android.os.Bundle;
+import android.util.Log;
+import android.view.MotionEvent;
 import android.view.View;
 import android.view.WindowManager;
 import android.widget.Button;
 import android.widget.ImageButton;
+import android.widget.ImageView;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import java.lang.reflect.Array;
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.HashSet;
 import java.util.Iterator;
+import java.util.LinkedHashSet;
+import java.util.List;
+import java.util.ListIterator;
 import java.util.Set;
 
 public class WrongActivity extends AppCompatActivity {
@@ -21,20 +30,36 @@ public class WrongActivity extends AppCompatActivity {
     private TextView chinaText, wordText, englishText;
     private SharedPreferences sharedPreferences;
     private WordHelper wordHelper;
+    private ImageView playVoice;
     Iterator it;
+    float x1 = 0;
+    float y1 = 0;
+    float x2 = 0;
+    float y2 = 0;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        getSupportActionBar().hide();
-        getWindow().setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN
-                , WindowManager.LayoutParams.FLAG_FULLSCREEN);
-
         setContentView(R.layout.activity_wrong);
         initControl();
         nextWrong();
     }
 
+    @Override
+    public boolean onTouchEvent(MotionEvent event) {
+        if (event.getAction() == MotionEvent.ACTION_DOWN) {
+            x1 = event.getX();
+            y1 = event.getY();
+        }
+        if (event.getAction() == MotionEvent.ACTION_UP) {
+            x2 = event.getX();
+            y2 = event.getY();
+            if(x2 - x1 > 200){
+                nextWrong();
+            }
+        }
+        return super.onTouchEvent(event);
+    }
 
     private void initControl(){
         chinaText = findViewById(R.id.china_text);
@@ -42,16 +67,21 @@ public class WrongActivity extends AppCompatActivity {
         englishText = findViewById(R.id.english_text);
         nextWrong = findViewById(R.id.i_know_btn);
         backBtn = findViewById(R.id.back_btn);
-
+        playVoice = findViewById(R.id.play_voice);
         wordHelper = new WordHelper(this);
         sharedPreferences = getSharedPreferences("config",MODE_PRIVATE);
-        Set<String> wrong= sharedPreferences.getStringSet("wrong",new HashSet<>());
+        Set<String> wrong= sharedPreferences.getStringSet("wrong",new LinkedHashSet<>());
         it = wrong.iterator();
 
         nextWrong.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                nextWrong();
+                try{
+                    it.remove();
+                    nextWrong();
+                }catch (IllegalStateException e){
+                    setFinalText();
+                }
             }
         });
         backBtn.setOnClickListener(new View.OnClickListener() {
@@ -66,18 +96,23 @@ public class WrongActivity extends AppCompatActivity {
         chinaText.setText(word.getChinese());
         wordText.setText(word.getWord());
         englishText.setText(word.getSoundmark());
+        playVoice.setVisibility(View.VISIBLE);
+    }
+
+    private void setFinalText(){
+        chinaText.setText("你真棒！");
+        wordText.setText("");
+        englishText.setText("");
+        playVoice.setVisibility(View.INVISIBLE);
     }
 
     private void nextWrong(){
         if(it.hasNext()){
-            String index = (String) it.next();
-            Word word = wordHelper.getXWord(Integer.parseInt(index));
+            String wordIndex = (String) it.next();
+            Word word = wordHelper.getXWord(Integer.parseInt(wordIndex));
             setText(word);
         }else{
-            Toast.makeText(this,"天哪，你好厉害",Toast.LENGTH_SHORT).show();
+            setFinalText();
         }
     }
-
-
-
 }